split image in separate types
