+++ pixman/examples/alpha.rs
@@ -1,4 +1,6 @@
-use pixman::{Fixed, FormatCode, GradientStop, Image, Operation, Point, Repeat, Transform};
+use pixman::{
+    Fixed, FormatCode, GradientStop, Image, LinearGradient, Operation, Point, Repeat, Transform,
+};
 
 const WIDTH: usize = 400;
 const HEIGHT: usize = 200;
@@ -19,7 +21,7 @@ pub fn main() {
     ]);
 
     let mut alpha = [0x4f00004fu32; WIDTH * HEIGHT]; /* pale blue */
-    let mut alpha_img = Image::from_bits(
+    let alpha_img = Image::from_slice_mut(
         FormatCode::A8R8G8B8,
         WIDTH,
         HEIGHT,
@@ -30,7 +32,7 @@ pub fn main() {
     .unwrap();
 
     let mut dest = [0xffffff00u32; WIDTH * HEIGHT]; /* yellow */
-    let mut dest_img = Image::from_bits(
+    let dest_img = Image::from_slice_mut(
         FormatCode::A8R8G8B8,
         WIDTH,
         HEIGHT,
@@ -41,17 +43,10 @@ pub fn main() {
     .unwrap();
 
     let mut src = [0xffff0000; WIDTH * HEIGHT];
-    let mut src_img = Image::from_bits(
-        FormatCode::A8R8G8B8,
-        WIDTH,
-        HEIGHT,
-        &mut src,
-        WIDTH * 4,
-        false,
-    )
-    .unwrap();
+    let src_img =
+        Image::from_slice(FormatCode::A8R8G8B8, WIDTH, HEIGHT, &mut src, WIDTH * 4).unwrap();
 
-    let mut grad_img = Image::linear_gradient(p1, p2, &stops).unwrap();
+    let grad_img = LinearGradient::new(p1, p2, &stops).unwrap();
     grad_img.set_transform(transform).unwrap();
     grad_img.set_repeat(Repeat::Pad);
 
@@ -69,7 +64,7 @@ pub fn main() {
         HEIGHT as u16,
     );
 
-    src_img.set_alpha_map(Some(&alpha_img), 10, 10);
+    let src_img = src_img.set_alpha_map(&alpha_img, 10, 10);
 
     dest_img.composite(
         Operation::Over,
@@ -85,7 +80,7 @@ pub fn main() {
         HEIGHT as u16,
     );
 
-    let mut out_img = Image::new(
+    let out_img = Image::new(
         FormatCode::A8B8G8R8,
         dest_img.width(),
         dest_img.height(),
